refactor: member query 관련 DTO 패키지 분리
