+++ src/main/java/com/umc/product/member/application/port/in/query/dto/MemberProfileInfo.java
@@ -1,4 +1,4 @@
-package com.umc.product.member.application.port.in.query;
+package com.umc.product.member.application.port.in.query.dto;
 
 import com.umc.product.member.domain.MemberProfile;
 import lombok.Builder;
